New Logging and Log Parsing Routines
We added tools to parse the log files generated by our experiments.
The parser will automatically apply sanity tests and allows for
extracting most of the interesting information from the log files.
It is already used by one small tool class which generates an end-result
table CSV file from all the data generated by an experiment.
In order to facilitate that, the logging methods have been consolidated
and overhauled.
This has led to some changes in the way algorithm parameters are stored.
Apart from that, the logs of the previous versions are compatible to the
new logging facilities.