begin to use checkstyle
We now use the checkstyle plugin in our Maven builds.
The goal is to overall improve the code quality, but specifically to
step-by-step find the violations of the Java naming convention, most
specifically names containing underscores "_".
These become discouraged with Java 9.
If eventually we want to go to Java 9, 11, or maybe 17, then we should
fix these violations.
Right now, most filters for checkstyle are turned off.
Eventually, I will turn them on step-by-step and then fix the errors as
they occur.
Eventually, the code will be in a better state.
After reaching such state, migration to a newer Java version might be
considered (or not).