API overhaul: Search Operators no longer Part of IBlackBoxProcess
I moved the search operators out of IBlackBoxProcess.
The reason is that many algorithms do not make use of some of them.
For instance, the EDAs only need a nullary search operation, the rest
is done by the model.
From the semantics, it might be better and clearer to store the
operators as member variables of the algorithm implementation.
They are part of the algorithm configuration.
Thus, semantically, they belong next to the other algorithm parameters.
This also makes it easier for the programmer or reader to understand
what is part of the optimization problem and what is part of the
algorithm setup.

Nevertheless, it might but subject to debate whether the searcher
operators are part of the representation of a problem or part of an
algorithm.
One might as well consider them a part of the problem representation.
This was the original reason to put them into IBlackBoxProcess.
Then again, the models used in an EDA could also be considered as a part
of the problem representation then.
But it might not be a good idea to add all sorts of things that may or
may not be part of the problem representation...

The search and solution space objects are still part of the black box
process. They allow us to copy and verify points in the search and
solution space. They must be part of the black box process, because
otherwise our method of automatically remembering the best results
cannot work.

...so, in summary, I decided to move the search operators out of the
black box processes and as member variables into the algorithms.