Major Change to the Experimentation API: Algorithms Instantiated per Run
We now instantiate the algorithms on per-run basis, no longer on a
per-problem basis.
The reason is that if we use the same instance for multiple runs, there
is a possibility that due to a programming issue, the runs are not
independent.
Such issues might be quite intricate.
Instantiation per-run is wasteful, indeed.
But the experimentation API is for experiments, not for productive use.
Thus, I will favor correctness and peace of mind over efficiency.