Tree Search Operators under-the-hood Improvements
The tree search operators have been improved under the hood.
Now, whenever a path through a tree is constructed, all nodes have the
same probability to be selected as path end.
Paths are constructed, e.g., when applying mutation or tree crossover to
select the nodes to be replaced.

Before this change, nodes closer to the root of the tree had a higher
replacement probability.
This means that larger changes to the tree were more likely.
With the new selection method, it is more likely that the leaves of the
tree are modified.
This should lead to a) smaller changes and hopefully b) to a better
balance between search steps that alter the tree very much and such that
only alter it slightly.